requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.0
openai==1.12.0
pydantic==2.6.1
requests==2.31.0
